Split firebase_functions_test.dart
