+++ lib/helpers/child_category_detailed_view_helper.dart
@@ -4,7 +4,7 @@ import 'package:focused_menu/focused_menu.dart';
 import 'package:seg_coursework_app/models/category_item.dart';
 import 'package:audioplayers/audioplayers.dart';
 import 'package:seg_coursework_app/models/image_details.dart';
-import 'package:seg_coursework_app/widgets/categoryItem/image_square.dart';
+import 'package:seg_coursework_app/widgets/general/image_square.dart';
 import 'package:seg_coursework_app/widgets/child_view/item_unavailable.dart';
 
 // Returns the grid of images, clickable and unavailable
@@ -13,6 +13,8 @@ Expanded getMainImages(List<CategoryItem> images) {
     child: LayoutBuilder(
       builder: (context, constraints) {
         final isLandscape = constraints.maxWidth > constraints.maxHeight;
+        final imageSize = isLandscape? constraints.maxHeight : constraints.maxWidth;
+
         return GridView.builder(
             key: const Key("mainGridOfPictures"),
             padding: const EdgeInsets.all(8.0),
@@ -20,7 +22,7 @@ Expanded getMainImages(List<CategoryItem> images) {
             gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
                 crossAxisCount: isLandscape? 5 : 3, childAspectRatio: 4 / 3),
             itemBuilder: (context, index) {
-              return getImage(index, images);
+              return getImage(images[index], imageSize);
             });
       }
     ),
@@ -46,37 +48,54 @@ Future speak(String text) async {
 }
 
 // Returns one image that on click blurs the background and is available
-Padding getImage(int index, List<CategoryItem> images) {
+Padding getImage(CategoryItem image, double imageSize) {
   return Padding(
     padding: const EdgeInsets.all(10.0),
     // Focused menu holder gives the blured background on click
-    child: images[index].availability
-        ? getAvailableItem(images, index)
-        : getUnavailableItem(images, index),
+    child: image.availability
+        ? getAvailableItem(image: image, imageSize: imageSize)
+        : getUnavailableItem(image),
   );
 }
 
 // Prepare the menu holder for available item to blur the background on click
-FocusedMenuHolder getAvailableItem(List<CategoryItem> images, int index) {
-  return FocusedMenuHolder(
-      openWithTap: true,
-      onPressed: () {
-        speak(images[index].name);
-      },
-      menuItems: const [],
-      blurSize: 5.0,
-      menuItemExtent: 45,
-      blurBackgroundColor: Colors.black54,
-      duration: const Duration(milliseconds: 100),
-      // gets image ImageSquare and adds a focused menu holder for effects
-      child: ImageSquare(
-          image: ImageDetails(
-              name: images[index].name, imageUrl: images[index].imageUrl)));
+class getAvailableItem extends StatelessWidget {
+  const getAvailableItem({
+    super.key,
+    required this.image, required this.imageSize
+  });
+
+  final CategoryItem image;
+  final double imageSize;
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+        onTap: () {
+          speak(image.name);
+          showDialog(context: context, useSafeArea: false ,builder: (context)
+          {
+            return AlertDialog(
+              content: ImageSquare(
+                height: imageSize,
+                width: imageSize,
+                image: ImageDetails(
+                  name: image.name, 
+                  imageUrl: image.imageUrl
+                )
+              ),
+            );
+          });
+        },
+        // gets image ImageSquare and adds a focused menu holder for effects
+        child: ImageSquare(
+            image: ImageDetails(
+                name: image.name, imageUrl: image.imageUrl)));
+  }
 }
 
 // Get image as an ImageSquare without any click effects,
 // if not available image is blurred and has unavailable icon
-FocusedMenuHolder getUnavailableItem(List<CategoryItem> images, int index) {
+FocusedMenuHolder getUnavailableItem(CategoryItem image) {
   return FocusedMenuHolder(
     onPressed: () {
       final player = AudioPlayer();
@@ -89,10 +108,10 @@ FocusedMenuHolder getUnavailableItem(List<CategoryItem> images, int index) {
             height: 200,
             width: 250,
             image: ImageDetails(
-                name: images[index].name, imageUrl: images[index].imageUrl)),
+                name: image.name, imageUrl: image.imageUrl)),
         Container(
           // if item is unavailable blur image and display icon, method imported from item_unavailable widget
-          child: images[index].availability ? null : makeUnavailable(),
+          child: image.availability ? null : makeUnavailable(),
         ),
       ],
     ),
