+++ lib/pages/child/child_main_menu.dart
@@ -12,7 +12,7 @@ import 'package:seg_coursework_app/services/auth.dart';
 import 'package:seg_coursework_app/services/check_connection.dart';
 import 'package:seg_coursework_app/widgets/child_view/logout_icon_button.dart';
 import 'package:seg_coursework_app/widgets/loading_indicators/custom_loading_indicator.dart';
-import 'customizable_row.dart';
+import 'child_menu_category_row.dart';
 
 /// Shows the list of categories available for the child to select from
 class ChildMainMenu extends StatefulWidget {
