+++ lib/widgets/timetable/timetable_row.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import '../../models/timetable.dart';
-import '../categoryItem/image_square.dart';
+import '../general/image_square.dart';
 
 ///This widget builds a timetable but without the arrows to be shown in the all saved timetables page.
 ///Not to be confused with TimetableList which is shown in the main visual timetable page.
